[refactor] webconfig 이용
