+++ src/main/java/com/wayble/server/ServerApplication.java
@@ -1,6 +1,7 @@
 package com.wayble.server;
 
 import com.wayble.server.direction.external.tmap.TMapProperties;
+import com.wayble.server.direction.external.kric.KricProperties;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.autoconfigure.data.elasticsearch.ReactiveElasticsearchRepositoriesAutoConfiguration;
@@ -16,7 +17,7 @@ import org.springframework.scheduling.annotation.EnableScheduling;
 @EnableJpaAuditing
 @EnableScheduling
 @EnableElasticsearchRepositories(basePackages = "com.wayble.server.explore.repository")
-@EnableConfigurationProperties(TMapProperties.class)
+@EnableConfigurationProperties({TMapProperties.class, KricProperties.class})
 @EntityScan(basePackages = "com.wayble.server")
 public class ServerApplication {
