+++ src/main/java/com/wayble/server/direction/external/kric/KricProperties.java
@@ -1,0 +1,10 @@
+package com.wayble.server.direction.external.kric;
+
+import org.springframework.boot.context.properties.ConfigurationProperties;
+
+@ConfigurationProperties(prefix = "kric.api")
+public record KricProperties(
+        String key,
+        String baseUrl
+) {
+}
