+++ src/main/java/com/wayble/server/direction/external/kric/dto/KricToiletRawBody.java
@@ -1,4 +1,4 @@
-package com.wayble.server.direction.dto.toilet;
+package com.wayble.server.direction.external.kric.dto;
 
 import java.util.List;
